classical-ml/requirements.txt
pyarrow==12.0.1
numpy==1.24.3
pandas==1.5.3
transformers==4.31.0
pyrecdp==0.1.5
scikit-learn==1.2.2
scikit-learn-intelex==2023.1.1
threadpoolctl==3.1.0
category_encoders==2.6.1
ray[tune]==2.2.0
xgboost==1.7.5
xgboost-ray==0.1.15
optuna==3.2.0
sigopt==8.8.2
pyyaml==6.0.1
raydp==1.5.0
daal4py==2023.1.1
simplejson==3.19.1
modin[ray]==0.22.2
pydantic==1.10.7